chore: Create utility module for benches
