+++ benches/basic.rs
@@ -1,16 +1,8 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use std::time::{Duration, Instant, SystemTime};
+use std::time::{Duration, SystemTime};
 
-fn bencher<I: Copy, O>(s: impl Fn() -> I, f: impl Fn(I) -> O) -> impl Fn(u64) -> Duration {
-    move |n| {
-        let v = s();
-        let now = Instant::now();
-        for _ in 0..n {
-            let _ = black_box(f(v));
-        }
-        now.elapsed()
-    }
-}
+mod util;
+use util::bencher;
 
 fn rand_year() -> i32 {
     fastrand::i32(datealgo::YEAR_MIN..=datealgo::YEAR_MAX)
@@ -77,6 +69,9 @@ fn rand_iwd() -> (i32, u8, u8) {
 }
 
 fn bench_basic(c: &mut Criterion) {
+    c.bench_function("overhead", |b| {
+        b.iter_custom(bencher(rand_date, |d| black_box(d)));
+    });
     c.bench_function("rd_to_date", |b| {
         b.iter_custom(bencher(rand_rd, |rd| datealgo::rd_to_date(black_box(rd))))
     });
